Move C++11 workarounds into separate header
